Separate event normalization into a separate module (+0 B)
