+++ compat/src/index.js
@@ -28,6 +28,7 @@ import { memo } from './memo';
 import { forwardRef } from './forwardRef';
 import { PureComponent } from './PureComponent';
 import { Children } from './Children';
+import { applyEventNormalization } from './events';
 
 const version = '16.8.0'; // trick libraries to think we are react
 
@@ -96,43 +97,6 @@ function cloneElement(element) {
  */
 function isValidElement(element) {
 	return !!element && element.$$typeof === REACT_ELEMENT_TYPE;
-}
-
-/**
- * Normalize event handlers like react does. Most famously it uses `onChange` for any input element.
- * @param {import('./internal').VNode} vnode The vnode to normalize events on
- */
-function applyEventNormalization({ type, props }) {
-	if (!props || typeof type != 'string') return;
-	let newProps = {};
-
-	for (let i in props) {
-		if (/^on(Ani|Tra)/.test(i)) {
-			props[i.toLowerCase()] = props[i];
-			delete props[i];
-		}
-		newProps[i.toLowerCase()] = i;
-	}
-	if (newProps.ondoubleclick) {
-		props.ondblclick = props[newProps.ondoubleclick];
-		delete props[newProps.ondoubleclick];
-	}
-	if (newProps.onbeforeinput) {
-		props.onbeforeinput = props[newProps.onbeforeinput];
-		delete props[newProps.onbeforeinput];
-	}
-	// for *textual inputs* (incl textarea), normalize `onChange` -> `onInput`:
-	if (
-		newProps.onchange &&
-		(type === 'textarea' ||
-			(type.toLowerCase() === 'input' && !/^fil|che|ra/i.test(props.type)))
-	) {
-		let normalized = newProps.oninput || 'oninput';
-		if (!props[normalized]) {
-			props[normalized] = props[newProps.onchange];
-			delete props[newProps.onchange];
-		}
-	}
 }
 
 /**
